// app/ts_types/MarketplaceBase.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@1.12.1.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export interface InstantiateMsg {
  admin: string
  collections: Collection[]
}
export interface Collection {
  address: string
  config: CollectionConfig
}
export interface CollectionConfig {
  contract_address: Addr
  royalty_fee_bps: number
  royalty_fee_recipient: Addr
  sell_denoms: string[]
}
export type ExecuteMsg =
  | {
      buy: {
        collection: string
        token_id: string
      }
    }
  | {
      unlist: {
        collection: string
        token_id: string
      }
    }
  | {
      list: {
        collection: string
        price: Coin
        token_id: string
      }
    }
  | {
      add_or_update_collection: {
        collection_address: string
        config: CollectionConfig
      }
    }
  | {
      remove_collection: {
        collection: string
      }
    }
  | {
      propose_new_admin: {
        new_admin?: string | null
      }
    }
  | {
      claim_admin_role: {}
    }
export type Uint128 = string
export interface Coin {
  amount: Uint128
  denom: string
}
export type QueryMsg =
  | {
      listing: {
        collection: string
        token_id: string
      }
    }
  | {
      listings: {
        limit?: number | null
        start_after?: number | null
      }
    }
  | {
      listings_by_owner: {
        limit?: number | null
        owner: string
        start_after?: number | null
      }
    }
  | {
      listings_by_collection: {
        collection: string
        limit?: number | null
        start_after?: number | null
      }
    }
  | {
      whitelisted_collections: {}
    }
  | {
      events: {
        collection: string
        limit?: number | null
        start_after?: number | null
        token_id: string
      }
    }
export type EventAction = "List" | "Buy" | "Unlist"
export interface EventsResponse {
  events: Event[]
}
export interface Event {
  action: EventAction
  metadata: true
  timestamp_nanos: number
}
export type Addr = string
export interface ListingResponse {
  listing: Listing
}
export interface Listing {
  collection: Addr
  listing_id: number
  price: Coin
  seller: Addr
  token_id: string
}
export interface ListingsResponse {
  listings: Listing[]
}
export interface ListingsByCollectionResponse {
  listings: Listing[]
}
export interface ListingsByOwnerResponse {
  listings: Listing[]
}
export interface WhitelistedCollectionsResponse {
  collections: CollectionConfig[]
}